requirements.txt
streamlit
opencv-python
numpy
Pillow
python-dotenv
Jinja2
weasyprint
openai

